requirements.txt
python-telegram-bot[job-queue]==20.6
requests
psycopg2-binary
python-dotenv
